converted return value from int to long to prevent overflowing

+++ core/src/main/java/dev/nandi0813/practice/manager/gui/guis/queue/QueueSelectorGui.java
@@ -37,7 +37,7 @@ public abstract class QueueSelectorGui extends GUI {
         build();
     }
 
-    protected abstract int getUpdateCooldownMinutes();
+    protected abstract long getUpdateCooldownMinutes();
 
     protected abstract String getQueueConfigPath();
