+++ core/src/main/java/dev/nandi0813/practice/manager/gui/guis/queue/UnrankedGui.java
@@ -14,7 +14,7 @@ public class UnrankedGui extends QueueSelectorGui {
     }
 
     @Override
-    protected int getUpdateCooldownMinutes() {
+    protected long getUpdateCooldownMinutes() {
         return ConfigManager.getInt("QUEUE.UNRANKED.GUI-UPDATE-MINUTE");
     }
